Renamed python package src to fui_kk
src is a bad name, and bad for search/grep
fui_kk is more useful, and follows python convention
