+++ fui_kk/old/plot-courses-old.py
