+++ fui_kk/plot_courses.py
