+++ tests/test_import.py
@@ -3,8 +3,8 @@
 import path_fix
 
 # Test that all modules can be imported:
-from src import *
+from fui_kk import *
 
 def test_import():
-    # Dummy function, actual test is "from src import *"
+    # Dummy function, actual test is "from fui_kk import *"
     pass
